+++ src/net_barQuasiStatic_wall_ini.cpp
@@ -492,7 +492,7 @@ void net_barQuasiStatic::wall_ini(lexer *p, fdm *a, ghostcell *pgc)
     }
 
     // Geometrical constraints for inner meshes, positiv clockwise (first nd meshes are boundary meshes)
-    for (int i = nd; i < meshID.size(); i++)
+    for (size_t i = nd; i < meshID.size(); i++)
     {
         for (int j = 0; j < nf; j++)
         {
